feat: update php-resilient-logger to 2.0
