+++ composer.json
@@ -9,7 +9,7 @@
     "version": "2024.16",
     "require": {
         "ext-json": "*",
-        "city-of-helsinki/php-resilient-logger": "^1.0",
+        "city-of-helsinki/php-resilient-logger": "^2.0",
         "composer/installers": "^2.0",
         "cweagans/composer-patches": "^1.6.7",
         "doctrine/lexer": "^2",
